+++ setup.py
@@ -3,7 +3,7 @@ import os
 from setuptools import find_packages, setup
 
 project = "hub"
-VERSION = "1.1"
+VERSION = "1.1.1"
 
 this_directory = os.path.abspath(os.path.dirname(__file__))
 with open(os.path.join(this_directory, "README.md")) as f:
